+++ assets/javascripts/kitten/karl/buttons/social-button-icon-words.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import {
+  FacebookButtonIconWords,
+  TwitterButtonIconWords,
+  LinkedinButtonIconWords,
+} from 'kitten/components/buttons/social-button-icon-words'
+
+export const KarlFacebookButtonIconWords = props => (
+  <FacebookButtonIconWords
+    className="k-u-margin-bottom-single"
+  >
+    Continuer avec Facebook
+  </FacebookButtonIconWords>
+)
+
+export const KarlTwitterButtonIconWords = props => (
+  <TwitterButtonIconWords
+    className="k-u-margin-bottom-single"
+  >
+    Continuer avec Twitter
+  </TwitterButtonIconWords>
+)
+
+export const KarlLinkedinButtonIconWords = props => (
+  <LinkedinButtonIconWords>
+    Continuer avec LinkedIn
+  </LinkedinButtonIconWords>
+)