+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Enable button type attribute override on `TextInputWithButton`.
+
 ## [19.4.0] - 2018-02-02
 
 Features: