+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Enable button type attribute override on `TextInputWithButton`.
+## [19.4.1] - 2018-02-07
+
+Fix:
+- Enable button type attribute override on `TextInputWithButton`.
 
 ## [19.4.0] - 2018-02-02